refactor: app 패키지를 의존하지 않는 ManualHandlerAdapter의 클래스명을 수정하고 spring 패키지로 이동
